[Move] #11 - DTO 그룹 생성 및 파일 이동
